pyproject.toml
[project]
name = "audio-agent-dojo"
version = "0.1.0"
description = "Voice AI experimentation playground - STT, TTS, orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
